Add task for fetching latest news from NHK.
This can be useful when developing offline, or under poor network conditions.

Since we stub all network requests in our tests to use this fetched data,
this file doesn't have to be updated everytime we run our tests.

Also, there were some typos in the README that are now fixed.